+++ solana/programs/matching-engine/tests/shimful/mod.rs
@@ -1,3 +1,6 @@
-pub mod shims;
+pub mod fast_market_order_shim;
+pub mod post_message;
 pub mod shims_execute_order;
+pub mod shims_make_offer;
 pub mod shims_prepare_order_response;
+pub mod verify_shim;
